# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Moj Prostor API",
        "description": "Moj Prostor je aplikacija koja omogućava pronalaženje različitih prostora i označavanje istih na Google mapi, kako bi ih drugi korisnici mogli pregledati i oceniti.",
        "termsOfService": "https://moj-prostor.rs/uslovi-koriscenja",
        "contact": {
            "name": "Support",
            "email": "api@moj-prostor.rs"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://api.localhost/v1"
        }
    ],
    "paths": {
...
        "/space/{spaceUid}/review/{spaceReviewUid}": {
            "get": {
                "tags": [
                    "SpaceReviews"
                ],
                "summary": "Fetch Space Review.",
                "description": "Used to Fetch Space Review.",
                "operationId": "FetchSpaceReviewController",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/spaceUid"
                    },
                    {
                        "$ref": "#/components/parameters/spaceReviewUid"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CategoryDto"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorDto"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorDto"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorDto"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorDto"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "SpaceReviews"
                ],
                "summary": "Delete Space Review.",
                "description": "Used to Delete Space Review.",
                "operationId": "DeleteSpaceReviewController",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/spaceUid"
                    },
                    {
                        "$ref": "#/components/parameters/spaceReviewUid"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorDto"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorDto"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorDto"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorDto"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "SpaceReviews"
                ],
                "summary": "Update Space Review.",
                "description": "Used to Update Space Review.",
                "operationId": "UpdateSpaceReviewController",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/spaceUid"
                    },
                    {
                        "$ref": "#/components/parameters/spaceReviewUid"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SpaceDto"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SpaceDto"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorDto"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorDto"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorDto"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorDto"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorDto"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/space/{spaceUid}/review": {
            "get": {
                "tags": [
                    "SpaceReviews"
                ],
                "summary": "Fetch All Space Reviews.",
                "description": "Used to Fetch All Space Reviews.",
                "operationId": "FetchAllSpaceReviewsController",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/spaceUid"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SpaceReviewsDto"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorDto"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorDto"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorDto"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "SpaceReviews"
                ],
                "summary": "Store New Space Review.",
                "description": "Used to Store New Space Review.",
                "operationId": "StoreSpaceReviewController",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/spaceUid"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SpaceReviewDto"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "headers": {
                            "Location": {
                                "description": "Location of the newly created space review.",
                                "schema": {
                                    "type": "string",
                                    "format": "uri"
                                }
                            }
                        },
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SpaceReviewDto"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorDto"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorDto"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorDto"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorDto"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "UserDto": {
                "title": "User",
                "required": [
                    "name",
                    "email",
                    "password"
                ],
                "properties": {
                    "uid": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "name": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string",
                        "format": "email"
                    },
                    "password": {
                        "type": "string",
                        "format": "password"
                    }
                },
                "type": "object"
            },
...
            "CategoryDto": {
                "title": "Category",
                "required": [
                    "name"
                ],
                "properties": {
                    "uid": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "name": {
                        "type": "string"
                    },
                    "numberOfReviews": {
                        "type": "integer",
                        "format": "int32"
                    }
                },
                "type": "object"
            },
...
            "SpaceDto": {
                "title": "Space",
                "required": [
                    "name",
                    "longitude",
                    "latitude"
                ],
                "properties": {
                    "uid": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "name": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "longitude": {
                        "type": "number",
                        "format": "double"
                    },
                    "latitude": {
                        "type": "number",
                        "format": "double"
                    },
                    "numberOfReviews": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "score": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "average": {
                        "type": "number",
                        "format": "double"
                    }
                },
                "type": "object"
            },
...
            "SpaceReviewDto": {
                "title": "SpaceReview",
                "required": [
                    "category",
                    "rate"
                ],
                "properties": {
                    "uid": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "user": {
                        "$ref": "#/components/schemas/UserDto"
                    },
                    "space": {
                        "$ref": "#/components/schemas/SpaceDto"
                    },
                    "category": {
                        "$ref": "#/components/schemas/CategoryDto"
                    },
                    "rate": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "description": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "SpaceReviewsDto": {
                "title": "SpaceReviews",
                "type": "array",
                "items": {
                    "$ref": "#/components/schemas/SpaceReviewDto"
                }
            },
            "ErrorDto": {
                "title": "Error",
                "required": [
                    "cid",
                    "timestamp",
                    "message"
                ],
                "properties": {
                    "cid": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "timestamp": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "message": {
                        "type": "string"
                    }
                },
                "type": "object"
            }
        },
        "parameters": {
...
            "spaceUid": {
                "name": "spaceUid",
                "in": "path",
                "description": "Used to identify space.",
                "required": true,
                "schema": {
                    "type": "string",
                    "format": "uuid"
                }
            },
            "spaceReviewUid": {
                "name": "spaceReviewUid",
                "in": "path",
                "description": "Used to identify space review.",
                "required": true,
                "schema": {
                    "type": "string",
                    "format": "uuid"
                }
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}